Fixed a an integral unsigned-to-sign conversion issue with a temporary long long work-around; templated functions were attempted but it caused some remote network federates to crash. Templated functions will be used once the data outgoing data corruption.
Also added the ability to map enumerated values to strings and resources rather than just enumeration constants. This allows for data file based enumerations via strings; and to map in certain files based on network integral value via resource properties.

+++ trunk/delta3d/inc/dtHLAGM/rprparametertranslator.h
@@ -152,8 +152,8 @@ namespace dtHLAGM
             const OneToManyMapping::ParameterDefinition& paramDef, 
             bool returnGameValue) const;
 
-         void SetIntegerValue(long value, dtGame::MessageParameter& parameter, const OneToManyMapping& mapping, unsigned parameterDefIndex) const;
-         long GetIntegerValue(const dtGame::MessageParameter& parameter, const OneToManyMapping& mapping, unsigned parameterDefIndex) const;
+         void SetIntegerValue(long long value, dtGame::MessageParameter& parameter, const OneToManyMapping& mapping, unsigned parameterDefIndex) const;
+         long long GetIntegerValue(const dtGame::MessageParameter& parameter, const OneToManyMapping& mapping, unsigned parameterDefIndex) const;
 
          void MapFromParamToWorldCoord(
             char* buffer,
